Final Art Piece
I drew a never ending knot. This particular design is a Buddhist design, that I have found to be associated with karma (which is my name and why the design was introduced to me). I think it represents culture because it is not only a cultural symbol, but also because its complicated and never ending: like culture.